+++ dimos/robot/assembly/zed_mid360/stereo_mount.urdf
